requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiohttp-socks==0.7.1
aiosignal==1.2.0
alembic==1.9.1
anyio==3.6.2
apparmor==3.1.2
appdirs==1.4.4
apsw==3.40.0.0
arabic-reshaper==2.1.3
asn1crypto==1.5.1
asttokens==2.0.8
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.1.0
autocommand==2.2.2
Automat==22.10.0
autotiling==1.8
backcall==0.2.0
beautifulsoup4==4.11.1
Brotli==1.0.9
brotlicffi==1.0.9.2
btrfsutil==6.1.3
cchardet==2.1.7
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
chatGPT @ git+https://github.com/mmabrouk/chatgpt-wrapper@f2d1101cb2f0e1b57fa9fef647a0eef04fea7e5c
click==8.1.3
click-log==0.4.0
cloudscraper==1.2.64
colorama==0.4.5
configobj==5.1.0.dev0
constantly==15.1.0
contourpy==1.0.6
cryptography==38.0.1
css-parser==1.0.8
cssselect==1.2.0
cssselect2==0.7.0
cum==0.9.1
cycler==0.11.0
Cython==0.29.33
debugpy==1.6.3
decorator==4.4.2
distro==1.8.0
dnspython==2.2.1
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.1.0
fastjsonschema==2.16.2
feedparser==6.0.10
flashfocus==2.2.4
fonttools==4.38.0
frozenlist==1.3.1
future==0.18.2
future-annotations==1.0.0
GeoIP==1.3.2
greenlet==2.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2020.1.16
html5-parser==0.4.10
html5lib==1.1
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
hyperlink==21.0.0
i3ipc==2.2.1
icalendar==5.0.4
idna==3.4
ifaddr==0.2.0
importlib-metadata==5.0.0
incremental==22.10.0
inflate64==0.3.1
inflect==6.0.2
iniconfig==1.1.1
ipykernel==6.16.1
ipython==8.5.0
jaraco.context==4.3.0
jaraco.functools==3.5.2
jaraco.text==3.11.1
jedi==0.18.1
jeepney==0.8.0
Jinja2==3.1.2
jsonpickle==3.0.1
jupyter_client==7.4.3
jupyter_core==4.11.2
khal==0.10.5
kiwisolver==1.4.4
LibAppArmor==3.1.2
libtorrent==2.0.8
lit==15.0.7.dev0
lxml==4.9.2
maigret==0.4.4
Mako==1.2.4
manjaro-sdk==0.1
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.19.0
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mechanize==0.4.8
mergedeep==1.3.4
mock==4.0.3
more-itertools==9.0.0
msgpack==1.0.4
multidict==6.0.2
multivolumefile==0.2.3
natsort==8.2.0
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==2.5.1
nftables==0.1
npyscreen==4.10.5
nspektr==0.4.0
numpy==1.23.3
nwg-wrapper==0.1.3
openai==0.26.5
ordered-set==4.1.0
oscrypto==1.3.0
outcome==1.2.0
packaging==23.0
pacman-mirrors==4.23.2
parso==0.8.3
pdftotext==2.2.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
playwright==1.30.0
pluggy==1.0.0
ply==3.11
praw==7.6.1
prawcore==2.3.0
prompt-toolkit==3.0.31
psutil==5.9.4
ptyprocess==0.7.0
pulsemixer==1.5.1
pure-eval==0.2.2
py==1.11.0
py7zr==0.20.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybcj==1.0.1
pycares==4.2.2
pychm==0.8.6
pycountry==22.3.5
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.12.0
pydantic==1.10.4
pyee==9.0.4
Pygments==2.14.0
PyGObject==3.42.2
pyHanko==0.14.0
pyhanko-certvalidator==0.19.5
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyPDF2==2.10.4
PyPDF3==1.0.6
pyppmd==1.0.0
PyQt6==6.4.2
PyQt6-sip==13.4.1
PyQt6-WebEngine==6.4.0
PySide6==6.4.2
PySocks==1.7.1
pytest==7.1.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-socks==2.1.1
python-telegram-bot==20.1
python-xlib==0.31
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pyvis==0.2.1
pyxdg==0.28
PyYAML==6.0
pyzmq==24.0.1
pyzstd==0.15.3
qrcode==7.3.1
ranger-fm==1.9.3
regex==2022.10.31
rencode==1.0.6
reportlab==3.6.11
requests==2.28.1
requests-futures==1.0.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==13.3.1
scipy==1.9.1
selenium==4.7.2
selenium-wire==2.1.2
service-identity==21.1.0
setproctitle==1.3.2
sgmllib3k==1.0.0
shiboken6==6.4.2
shiboken6-generator==6.4.2
shtab==1.5.8
six==1.16.0
sniffio==1.3.0
socid-extractor==0.0.23
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.20
stack-data==0.5.1
stem==1.8.0
svglib==1.4.1
systemd-python==235
team==1.0
termcolor==2.2.0
texttable==1.6.7
tinycss2==1.1.1
tokenize-rt==4.2.1
tomli==2.0.1
tornado==6.2
torrequest==0.1.0
tqdm==4.64.1
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
trove-classifiers==2023.2.9
Twisted==22.10.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
unrardll==0.1.5
update-checker==0.18.0
uritools==4.0.0
urllib3==1.26.13
urwid==2.1.2
user-agent==0.1.10
validate==5.1.0.dev0
validate-pyproject==0.12.1
vimgolf==0.4.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.5.0
win32-setctime==1.1.0
wsproto==1.2.0
xcffib==1.2.0
xhtml2pdf==0.2.8
XMind==1.2.0
xpybutil==0.0.6
yarl==1.7.2
zeroconf==0.39.4
zipp==3.13.0
zope.interface==5.5.2
zstandard==0.19.0
